Add Infoblox end to end test
